0.6.0 Part 2 (#186)
* Make changelog data driven, and more UI scaling improvements

* UI updates

* add extra null checks to redraw

* change a few default settings

* Enable actor name censoring by default

* Add better null checks in character revert logic

* Fix key binds not saving when changed

* Made updates to overlay (Thank you Caraxi!)

Some code from fork LivePose!

Thank you Caraxi!

https://github.com/Caraxi/LivePose/blob/69afd7ba4f46611ac6055266f2524d1ac1d22454/LivePose/UI/Windows/Specialized/PosingOverlayToolbarWindow.cs#L337

* Yep, doing this again

* Fixing this

* Fixing null handling in `CharacterHandlerService`

* Move camera widget into its own file and add new menu items

* Add `ToggleSelecterStrip`

* Add `ResetSelectedBone` and update Overlay

* Change default size of MainWindow

* Disable MCDF help command

* Add VieraEars visibility to AppearanceService

* Turns out I was smarter than I thought

I came here to fix this, but I fixed and didn't even know I fixed it!

fascinating

* various changes

* Optimize `CalculateClickables` and make gizmo show up even if all bones are disabled in the filter

- Still need to add a setting to make this disableable

- Thank you Ny for the suggestion

* Update kofi.txt

* This has been sitting here for so long I should just commit this already huh?

Dynamis

use it

* Fix color handling (Thank you Ny)

* Update TimeWidget.cs

* Update PosingGraphicalWindow.cs

* Oooo I can now add this one now to

* Isn't tested but I'll add it now

* Try to fix FreeCam spawning

* UI updates

* Add clipboard copy/paste for transform editor and UI updates

* This doesn't really fix this

* Add equipment buttons to popup

* Small UI tweaks

* Refactor camera and cutscene input handling

* Add keybind for freezing posing

* Add open `Camera Editor Window` MenuItem to `CameraContainer`

* Add invisible equipment button to gear editor

* Update MathHelpers.cs

* I will push this now so we get it out of the way

* oops

* Update changelog.yaml

* Fix build

* Really fix the build this time

+++ Brio/Game/Actor/CharacterHandlerService.cs
@@ -51,21 +51,22 @@ public class CharacterHandlerService : IDisposable
                 var character = _dalamudService.GetGposeCharacterFromObjectTableByName(entry.Name, onlyGposeCharacters: true);
                 if(character is null)
                 {
-                    CharacterHandler.Remove(entry);
                     RevertMCDF(entry).GetAwaiter().GetResult();
                 }
             }
+            CharacterHandler.Clear();
         }
     }
 
     public async Task RevertMCDF(CharacterHolder mCDFCharacterHolder)
     {
-        if(mCDFCharacterHolder.GameObject.Address == nint.Zero || mCDFCharacterHolder.Name.IsNullOrEmpty()) return;
+        if(mCDFCharacterHolder.GameObject is not null)
+            _glamourerService.UnlockAndRevertCharacter(mCDFCharacterHolder.GameObject);
 
-        _glamourerService.UnlockAndRevertCharacter(mCDFCharacterHolder.GameObject);
-        _glamourerService.UnlockAndRevertCharacterByName(mCDFCharacterHolder.Name);
+        if(mCDFCharacterHolder.Name.IsNullOrEmpty() is false)
+            _glamourerService.UnlockAndRevertCharacterByName(mCDFCharacterHolder.Name);
 
-        if(mCDFCharacterHolder.GameObject.Address != nint.Zero)
+        if(mCDFCharacterHolder.GameObject is not null && mCDFCharacterHolder.GameObject.Address != nint.Zero)
         {
             _customizePlusService.RemoveTemporaryProfile(mCDFCharacterHolder.GameObject);
             await _penumbraService.Redraw(mCDFCharacterHolder.GameObject, true).ConfigureAwait(false);
@@ -74,7 +75,7 @@ public class CharacterHandlerService : IDisposable
 
     public async Task Revert(IGameObject obj, bool afterGpose = false)
     {
-        if(obj.Address == nint.Zero) return;
+        if(obj is null) return;
 
         _glamourerService.UnlockAndRevertCharacterByName(obj.Name.TextValue);
         _glamourerService.UnlockAndRevertCharacter(obj);
